requirements.txt
supabase>=2.3.1
python-dotenv>=1.0.0
websockets>=12.0